backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
pymysql==1.1.1
python-multipart==0.0.9
requests==2.31.0
colorama==0.4.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.0
alembic==1.10.3
httpx==0.27.0
email-validator==2.1.2
